requirements.txt
gensim
nltk
polars
scipy
sentence-transformers
